+++ models/__init__.py
@@ -1,27 +1,11 @@
 from .build import build_model_from_cfg
-import models.TopNet
-import models.not_ours.PoinTr
-import models.GRNet
-import models.not_ours.PCN
-import models.not_ours.FoldingNet
-import models.not_ours.SnowFlakeNet
-import models.not_ours.AdaPoinTr
-# import models.ImgResNetEncDecAdaPoinTr
-# import models.ImgResNetEncAdaPoinTr
-import models.good.ImgResNetEncAdaPoinTrVariableLoss
-# import models.ImgResNetFoldNetAdaPoinTr
-import models.good.SegEncAdaPoinTr
-# import models.SegEncDecAdaPoinTr
-# import models.SegEncSimAdaPoinTr
-# import models.SegFoldNetAdaPoinTr
-# import models.ImgResNetEncSimAdaPoinTr
-# import models.ImgResNetEncAlphaAdaPoinTr
-# import models.SegCrossEncAdaPoinTr
-# import models.SegCrossDecAdaPoinTr
-# import models.ImgResNetСrossEncAdaPoinTr
-# import models.ImgResNet50EncAdaPoinTrVariableLoss
-# import models.AdaPoinTrVariableLoss
-# import models.ConvNextEncAdaPoinTrVariableLoss
-# import models.ConvNextEncAdaPoinTrVariableLossLin
-# import models.ImgResNet50EncAdaPoinTrVariableLossLin
-import models.good.ImgEncSegDecAdaPoinTrVariableLoss
+import models.external.TopNet
+import models.external.GRNet
+import models.external.PCN
+import models.external.FoldingNet
+import models.external.SnowFlakeNet
+import models.external.AdaPoinTr
+
+import models.internal.ImgResNetEncAdaPoinTrVariableLoss
+import models.internal.SegEncAdaPoinTr
+import models.internal.ImgEncSegDecAdaPoinTrVariableLoss
